latest/v/tt/SportsHub/requirements.txt
absl-py==2.0.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.1
comtypes==1.2.1
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.1
django-allauth==0.59.0
django-crispy-forms==2.1
django-jazzmin==2.6.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-tempus-dominus==5.1.2.17
django-tinymce==3.6.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.0
fqdn==1.5.1
idna==3.6
ipykernel==6.28.0
ipython==8.20.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.4
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mediapipe==0.10.9
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.14.1
nbformat==5.9.2
nest-asyncio==1.5.8
notebook_shim==0.2.3
numpy==1.26.3
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
overrides==7.4.0
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
pbr==6.0.0
pillow==10.2.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.7
pure-eval==0.2.2
pycparser==2.21
pygame==2.5.2
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pypiwin32==223
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-openid==3.2.0
pyttsx3==2.90
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
razorpay==1.4.1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
SpeechRecognition==3.10.1
sqlparse==0.4.4
stack-data==0.6.3
stevedore==5.1.0
terminado==0.18.0
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
urllib3==2.1.0
virtualenv==20.25.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
